试用账号申请添加重复密码 支付宝金额字段类型修改 人员模块 字段映射模块修改 授权界面提那家用户名

+++ src/main/java/cn/com/sinofaith/bean/zfbBean/ZfbZhmxJczzEntity.java
@@ -1,6 +1,7 @@
 package cn.com.sinofaith.bean.zfbBean;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "zfbzhmx_jczz")
@@ -10,9 +11,9 @@ public class ZfbZhmxJczzEntity {
     private String jymc;
     private long jyzcs;
     private long czzcs;
-    private double czzje;
+    private BigDecimal czzje;
     private long jzzcs;
-    private double jzzje;
+    private BigDecimal jzzje;
     private String insert_time;
     private long aj_id;
 
@@ -68,11 +69,11 @@ public class ZfbZhmxJczzEntity {
 
     @Basic
     @Column(name = "czzje",nullable = false,precision = 0)
-    public double getCzzje() {
+    public BigDecimal getCzzje() {
         return czzje;
     }
 
-    public void setCzzje(double czzje) {
+    public void setCzzje(BigDecimal czzje) {
         this.czzje = czzje;
     }
 
@@ -88,11 +89,11 @@ public class ZfbZhmxJczzEntity {
 
     @Basic
     @Column(name = "jzzje",nullable = false,precision = 0)
-    public double getJzzje() {
+    public BigDecimal getJzzje() {
         return jzzje;
     }
 
-    public void setJzzje(double jzzje) {
+    public void setJzzje(BigDecimal jzzje) {
         this.jzzje = jzzje;
     }
